requirements.txt
requests~=2.32.5
python-telegram-bot~=13.15
environs~=14.5.0
standard-imghdr==3.13.0
urllib3==1.26.15
